# services/bank/docs/swagger.yaml
basePath: /
definitions:
  accounts.accountResponse:
    properties:
      balance:
        example: 200
        type: integer
      currency:
        example: RUB
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      user_id:
        example: 450e8400-e29b-41d4-a716-446655440001
        type: string
    type: object
  accounts.moneyRequest:
    properties:
      amount:
        example: 50025
        type: integer
    type: object
  accounts.transferRequest:
    properties:
      amount:
        example: 50025
        type: integer
      to_account_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
    type: object
  transfers.OperationType:
    enum:
    - deposit
    - withdrawal
    - transfer
    type: string
    x-enum-varnames:
    - OperationDeposit
    - OperationWithdrawal
    - OperationTransfer
  transfers.Transfer:
    properties:
      amount:
        type: integer
      createdAt:
        type: string
      currency:
        type: string
      fromAccID:
        type: string
      id:
        type: string
      toAccID:
        type: string
      type:
        $ref: '#/definitions/transfers.OperationType'
    type: object
  users.createRequest:
    properties:
      email:
        example: ivan@example.com
        type: string
      name:
        example: Ivan
        type: string
      password:
        example: strong-password
        type: string
    type: object
  users.loginRequest:
    properties:
      email:
        example: ivan@example.com
        type: string
      password:
        example: strong-password
        type: string
    type: object
  users.userResponse:
    properties:
      created_at:
        example: "2026-08-27T12:00:00Z"
        type: string
      email:
        example: ivan@example.com
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      name:
        example: Ivan
        type: string
    type: object
info:
  contact: {}
  description: |-
    Учебный REST API банковского сервиса.
    Для защищённых endpoints сначала выполните POST /auth/login.
    Access JWT и rotating refresh token устанавливаются в HttpOnly cookies.
  title: Bank API
  version: "1.0"
paths:
  /accounts/deposit:
    post:
      consumes:
      - application/json
      description: Пополняет баланс аккаунта текущего пользователя и сохраняет финансовую
        операцию. Требуется access JWT cookie.
      parameters:
      - description: Сумма пополнения
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/accounts.moneyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/accounts.accountResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Пополнить баланс
      tags:
      - accounts
  /accounts/my:
    get:
      description: Возвращает аккаунт пользователя, определённого по JWT cookie.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/accounts.accountResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получить свой аккаунт по id
      tags:
      - accounts
  /accounts/withdraw:
    post:
      consumes:
      - application/json
      description: Списывает средства с баланса аккаунта текущего пользователя и сохраняет
        финансовую операцию. Требуется access JWT cookie.
      parameters:
      - description: Сумма списания
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/accounts.moneyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/accounts.accountResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Вывести средства
      tags:
      - accounts
  /auth/login:
    post:
      consumes:
      - application/json
      description: Проверяет email и пароль, создаёт сессию и устанавливает access/refresh
        токены в HttpOnly cookies.
      parameters:
      - description: Учётные данные
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/users.loginRequest'
      produces:
      - application/json
      responses:
        "204":
          description: Успешная аутентификация
          headers:
            Set-Cookie:
              description: HttpOnly cookies access_token и refresh_token
              type: string
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Войти
      tags:
      - auth
  /auth/logout:
    post:
      description: Отзывает текущую сессию по refresh cookie и удаляет обе auth cookies.
      responses:
        "204":
          description: Выход выполнен
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Выйти из текущей сессии
      tags:
      - auth
  /auth/logout-all:
    post:
      description: Отзывает все refresh-сессии пользователя и удаляет auth cookies
        текущего устройства. Требуется access JWT.
      responses:
        "204":
          description: Все сессии отозваны
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Выйти со всех устройств
      tags:
      - auth
  /auth/refresh:
    post:
      description: Проверяет и ротирует refresh token, затем устанавливает новую пару
        HttpOnly cookies.
      responses:
        "204":
          description: Токены обновлены
          headers:
            Set-Cookie:
              description: Новые HttpOnly cookies access_token и refresh_token
              type: string
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Обновить токены
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Создаёт пользователя и его банковский аккаунт в одной транзакции.
      parameters:
      - description: Данные для регистрации
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/users.createRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/users.userResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Зарегистрировать пользователя
      tags:
      - auth
  /transfers/history:
    get:
      description: Get transfer history for the current user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/transfers.Transfer'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get transfer history
      tags:
      - transfers
  /transfers/history/{id}:
    get:
      description: Get transfer history for the current user by ID
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/transfers.Transfer'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get transfer history by ID
      tags:
      - transfers
  /users/{id}:
    get:
      description: Возвращает профиль другого пользователя. Требуется JWT cookie.
      parameters:
      - description: UUID пользователя
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/users.userResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получить профиль по ID
      tags:
      - users
  /users/me:
    get:
      description: Возвращает профиль пользователя, определённого по JWT cookie.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/users.userResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получить свой профиль
      tags:
      - users
schemes:
- http
- https
swagger: "2.0"
tags:
- description: Регистрация и аутентификация.
  name: auth
- description: Профили пользователей. Требуется cookie access_token.
  name: users
- description: Банковский аккаунт, пополнение и списание средств. Требуется cookie
    access_token.
  name: accounts